Create frmPayment.cs
Created the payment form, but I'm not sure how it's supposed to look — in which case all the code for it is in frmPayment_Code.

Pass the student's id to frmPayment from whichever form displays frmPayment to complete the order.

Replace "column name" with the name of the column which has the student's balance.