Created Concept Menu Form and Cart Form
Concept menu form consists of two checkedListBoxes. The student checks the items they would like to order and then adds the items to their cart using the Add to Cart.

Cart form list the checked items from the concept menu.

Should you give the student the ability to make changes to their cart before payment?

Lastly, the student clicks on the Checkout button to pay for the order.